Blatt 04 Aufgabe 2: 	a2.py: alles einkommentiert, ein paar Achsenbeschriftungen korrigiert 	a2.cpp: alles wieder dabei, nichts mehr auskommentiert, Zeiten geändert 	Dpendulum.cpp: Startwerte korrigiert, ein bisschen aufgeräumt, Änderung ohne Test: Zeile 214 auskommentiert, das sollte da nicht sein, aber ich habs erstmal nur auskommentiert und nicht gelöscht, weil ich jetzt keine Lust habe das nochmal durchlaufen zu lassen. Also einfach rausschmeißen wenn es bei euch läuft, einkommentieren wenn nicht. 	poincare.cpp: damit haben wir nie was getan, da steht nichts besonderes drinn und wird von make auch nicht aufgerufen. --> removed Ich glaub das sollte mal unsere main werden, aber das ist ja jetzt unsere a2.cpp
To do:  siehe letzter Commit, noch einige Kleinigkeiten zu tun!
	guckt euch mal diesen Link an: http://pages.physics.cornell.edu/~myers/teaching/ComputationalMethods/ComputerExercises/DoublePendulum/doublependulum.html Da steht was über die lineare Interpolation, aber ich versteh nicht ganz wie die da drauf kommen:

"A simpler approach is to interpolate. Suppose that s1=sin(theta1) at time slice i and s2=sin(theta1) at time slice i+1, and that one of s1,s2 is positive and the other is negative. Suppose the values of theta2 at these two slices are θ1,θ2. If we assume a linear relation between s and θ then we should have θ=[(s-s1)θ2-(s-s2)θ1]/(s2-s1). In particular the value of θ when s=0 will be θ0=(s2θ1-s1θ2)/(s2-s1)."